+++ shared/shared.go
@@ -241,7 +241,8 @@ type TransactionCheckoutInfo struct {
 	// Entry mode of the payment details.
 	EntryMode *TransactionCheckoutInfoEntryMode `json:"entry_mode,omitempty"`
 	// Internal unique ID of the transaction on the SumUp platform.
-	InternalId *int `json:"internal_id,omitempty"`
+	// Format: int64
+	InternalId *int64 `json:"internal_id,omitempty"`
 	// Unique code of the registered merchant to whom the payment is made.
 	MerchantCode *string `json:"merchant_code,omitempty"`
 	// Amount of the tip (out of the total transaction amount).
